proyecto final

+++ js/carrrito.js
@@ -14,7 +14,7 @@ const botonComprar = document.querySelector("#acciones-comprar");
 function mostrarProductos() {
     if (productosCarrito && productosCarrito.length > 0) {
 
-        
+
 
         carritoVacio.classList.add("disabled");
         carritoContenedor.classList.remove("disabled");
@@ -24,9 +24,9 @@ function mostrarProductos() {
         carritoContenedor.innerHTML = "";
 
         productosCarrito.forEach(producto => {
-        const div = document.createElement("div");
-        div.classList.add("carrito-producto");
-        div.innerHTML = `
+            const div = document.createElement("div");
+            div.classList.add("carrito-producto");
+            div.innerHTML = `
             <img class="carrito-img" src="${producto.imagen}" alt="${producto.titulo}">
             <div class="carrito-procuto-nombre">
                 <small>Producto</small>
@@ -47,16 +47,16 @@ function mostrarProductos() {
                     <button id="${producto.id}" class="carrito-eliminar"><i class="bi bi-trash3"></i></button>
         `;
 
-        carritoContenedor.append(div);
+            carritoContenedor.append(div);
 
         })
 
 
     } else {
-    carritoVacio.classList.remove("disabled");
-    carritoContenedor.classList.add("disabled");
-    carritoAcciones.classList.add("disabled");
-    carritoComprado.classList.add("disabled");
+        carritoVacio.classList.remove("disabled");
+        carritoContenedor.classList.add("disabled");
+        carritoAcciones.classList.add("disabled");
+        carritoComprado.classList.add("disabled");
 
     }
 
@@ -77,10 +77,24 @@ function actualizarBotonesEliminar() {
 }
 
 function eliminarCarrito(e) {
+    Toastify({
+        text: "Producto eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+            background: "linear-gradient(to right, #165C90, #70CCCD)",
+            borderRadius: "2rem",
+            fontSize: ".90rem",
+        },
+        onClick: function () { } // Callback after click
+    }).showToast();
     const idBoton = e.currentTarget.id;
     const index = productosCarrito.findIndex(producto => producto.id === idBoton);
 
-    
+
     productosCarrito.splice(index, 1);
     mostrarProductos();
 
@@ -89,11 +103,26 @@ function eliminarCarrito(e) {
 
 botonVaciar.addEventListener("click", vaciarCarrito)
 function vaciarCarrito() {
-
-    productosCarrito.length = 0;
-    localStorage.setItem("productos-carrito", JSON.stringify(productosCarrito));
-    mostrarProductos();
-
+    Swal.fire({
+        title: "Seguro que quiere eliminar su carrito?",
+        html: `Se van a borrar ${productosCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos`,
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, deseo vaciarlo",
+        cancelButtonText: "Cancelar",
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosCarrito.length = 0;
+            localStorage.setItem("productos-carrito", JSON.stringify(productosCarrito));
+            mostrarProductos();
+            Swal.fire({
+                title: "Carrito eliminado",
+                icon: "success"
+            });
+        }
+    });
 
 }
 
@@ -108,7 +137,7 @@ function comprarCarrito() {
 
     productosCarrito.length = 0;
     localStorage.setItem("productos-carrito", JSON.stringify(productosCarrito));
-    
+
     carritoVacio.classList.add("disabled");
     carritoContenedor.classList.add("disabled");
     carritoAcciones.classList.add("disabled");
